TM-4808 Modified the logic that disassociates a person from a project to clear out additional references. Noticed a unimplemented method in the ExceptionUtil class so added that at the same time. Updated the account import to have confirmation when removing staff from project. Removed a number of debug log statements to start cleaning up things. Refactored a number of closures to methods in the AccountImport logic.
git-svn-id: https://opsvcs.tdsops.com/repos/tdstm/trunk@6813 fecba46a-3362-0410-80e5-8c76cdb905a0

+++ grails-app/services/ProjectService.groovy
@@ -1514,6 +1514,11 @@ class ProjectService {
 			teamCodes = [ teamCodes ]
 		}
 
+		if (teamCodes.contains('STAFF')) {
+			// If you are getting this exception you should look at the PersonService.removeFromProject method
+			throw new InvalidParamException('STAFF can not be removed by removeTeamMember method')
+		}
+		
 		// Remove person/team references in the MoveEventStaff table			
 		String mesQuery = """DELETE from MoveEventStaff mes where
 			person=:person and role.id in (:teams) and
